requirements.txt
Django==2.2
django-phone-field==1.7.0
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
